+++ examples/nextjs/app/page.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+	return (
+		<main>
+			<h1>Démo de la documentation avec nextJS</h1>
+		</main>
+	)
+}