+++ packages/react-ui/source/component/Accordion.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import styled, { css } from 'styled-components'
+import { css, styled } from 'styled-components'
 import { AccordionItem } from '../contexts'
 import { Arrow } from './icons'